pyproject.toml
[project]
name = "final-project"
version = "0.1.0"
description = "Document Chatbot with RAG using Streamlit, LangGraph, and FAISS"
authors = [
    {name = "Heri Santoso", email = "heri.santoso@example.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
license = {text = "MIT"}
keywords = ["chatbot", "rag", "document", "streamlit", "langchain", "faiss"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "cohere>=5.18.0",
    "faiss-cpu>=1.12.0",
    "langchain>=0.3.27",
    "langchain-cohere>=0.4.6",
    "langchain-community>=0.3.30",
    "langchain-groq>=0.3.8",
    "langgraph>=0.6.8",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pillow>=11.3.0",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.50.0",
    "typing-extensions>=4.15.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
